avoid double scroll bars
It wasn't obvious on all devices, but the way that the sticky header was previously implemented caused a case of nested-scroll-bars, especially on a tablet device.

This change removes the fact that the header sticks to the top of the page, but also fixes the nested scroll bars.

The sticky header broke the UX of linking directly to a header in an artible, so I have been considering removing it anyway.